# openapi/backend.yaml
openapi: 3.0.3
info:
  title: Carbon Cut API
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/license/mit
  description: Backend-owned HTTP contract. Coverage starts with active collectivity project operations.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/collectivity/projects/{projectSlug}/current-inventory/input:
    parameters:
      - $ref: "#/components/parameters/ProjectSlug"
    put:
      operationId: saveCollectivityInventoryInput
      summary: Save a current inventory draft
      tags: [Collectivity projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryInputRequest" }
      responses:
        "200":
          description: Saved inventory and compact project reference
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventorySaveResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectSlug:
      name: projectSlug
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
    ProjectSetup:
      type: object
      required: [name, slug, country, territory, referenceYear, inventoryYears, applicability]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        country: { type: string, pattern: "^[A-Z]{3}$" }
        territory: { type: string, minLength: 1 }
        referenceYear: { type: integer }
        inventoryYears:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: integer }
          description: Must include referenceYear.
        applicability:
          type: object
          required: [port, agriculture]
          properties:
            airport:
              { type: boolean, description: Optional on input; backend defaults it to false. }
            port: { type: boolean }
            agriculture: { type: boolean }
...
    Inventory:
      type: object
      required:
        [
          id,
          projectId,
          setupPayload,
          inventoryInput,
          status,
          lockedYears,
          latestCalculationRunId,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string }
        projectId: { type: string, nullable: true }
        setupPayload: { $ref: "#/components/schemas/ProjectSetup" }
        inventoryInput:
          type: object
          nullable: true
          additionalProperties: true
        status: { type: string, enum: [draft, calculated, outdated] }
        lockedYears:
          type: array
          items: { type: integer }
        latestCalculationRunId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    InventoryInputRequest:
      type: object
      required: [inventoryInput]
      properties:
        inventoryInput: { $ref: "#/components/schemas/JsonObject" }
    InventorySaveProject:
      type: object
      required: [id, slug, currentInventoryId, updatedAt]
      properties:
        id: { type: string }
        slug: { type: string }
        currentInventoryId: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    InventorySaveResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [project, currentInventory]
          properties:
            project: { $ref: "#/components/schemas/InventorySaveProject" }
            currentInventory: { $ref: "#/components/schemas/Inventory" }
    ErrorResponse:
      type: object
      required: [data, error]
      properties:
        data: { type: object, nullable: true }
        error:
          type: object
          required: [status, name, message, details]
          properties:
            status: { type: integer }
            name: { type: string }
            message: { type: string }
            details: { $ref: "#/components/schemas/JsonObject" }